feat(mlop-2760): renaming and removing unused test
